Adds a check to make sure $query is instance of WP_Query
The Enterprise Search feature calls methods in APC. APC will cause a fatal error with eterprise search when $query is not an instance of WP_Query.

This PR prevents that fatal error by making sure that it won't perform a $query is not a WP_Query. After this change, when the check fails, no fatal error is incurred and the search is successful.